% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-priors.R
\name{chain_prior_plot}
\alias{chain_prior_plot}
\title{Plots Prior Distributions}
\usage{
chain_prior_plot(
  interval_start_times = c(0, 3, 6, 12, 18),
  interval_end_times = c(3, 6, 12, 18, 24),
  doses = seq(0, 3),
  prior_list = NULL,
  prior_name = NULL
)
}
\arguments{
\item{interval_start_times}{numeric vector,
default is c(0,3,6,12,18).}

\item{interval_end_times}{numeric vector,
default is c(3,6,12,18,24)}

\item{doses}{numeric vector,
default is seq(0,3)}

\item{prior_list}{Named list with each object
containing the output from the chain_priors_from_data function.}

\item{prior_name}{string for prior_list object to
use.}
}
\value{
A named list with two ggplot objects: \code{time_v_prior} and
\code{timebins_v_prior}.
}
\description{
Plots Prior Distributions
}
\examples{
library(PriorRhythm)

SynHist <- system.file("extdata", "SyntheticHistorical",
  package = "PriorRhythm", mustWork = TRUE
)

Long_dat <- readRDS(file = file.path(SynHist, "Syn_long.RDS"))

# normal use
p_out <- Long_dat |>
  dplyr::filter(.data$study_code \%in\% c(1, 2)) |>
  chain_priors_from_data(
    treatment_col    = "treatment_code",
    study_col        = "study_code",
    animal_col       = "animal_code",
    period_col       = "period_code",
    assay_names_col  = "parameter",
    assay_name       = "assay1",
    alpha0_pop_mean = 250
  )

prior_plots <- chain_prior_plot(
  interval_start_times = c(0, 3, 6, 12, 18),
  interval_end_times   = c(3, 6, 12, 18, 24),
  doses                = seq(0, 3),
  prior_list           = p_out,
  prior_name           = "assay1"
)

cat("Plot names:", names(prior_plots), "\n")
print(prior_plots$time_v_prior)
print(prior_plots$timebins_v_prior)
}
