% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-config.R
\name{print.PriorRhythmConfig}
\alias{print.PriorRhythmConfig}
\title{Print a PriorRhythmConfig Object}
\usage{
\method{print}{PriorRhythmConfig}(x = NULL, ...)
}
\arguments{
\item{x}{a \code{PriorRhythmConfig} object.}

\item{...}{unused; present for S3 method compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a PriorRhythmConfig Object
}
